Export core package with OSS compliant version
Motivation:

In order to allow use of the dse-driver as a drop-in replacement for
cassandra-driver-core in an OSGi environment, the dse-driver needs to
export com.datastax.driver.core.* with a version of the OSS driver it is
compliant with.  Previously the bundle shared the same name as the core
bundle (com.datastax.driver.core) but used the version of the dse-driver
which is not the same as the OSS driver core module.

Modifications:

Renamed OSGi bundle from com.datastax.driver.core to
com.datastax.driver.dse.

Added property oss-driver.version for maintaining the compliant OSS
driver version.

Exported com.datastax.driver.core.* with oss-driver.version.

Updated OSGi tests to use the dse driver and grab version dependencies
from the maven pom instead of hard coding.

Result:

dse-driver can now be used as a drop-in replacement for
cassandra-driver-core in an OSGi environment.