Move DSE configuration logic into CCMBridge
Motivation:

CCMBridge.Builder previously didn't not account for that fact that the
user could be running with DSE enabled by default.  In this case, there
needs to be a mechanism for disabling DSE for certain tests.

In addition, there was some logic for DSE 5.0+ specific clusters that
was in CCMDseTestsSupport.  This has been moved into CCMBridge.Builder
so it may be more globally applied.

Modifications:

Move logic from CCMDseTestsSupport to CCMBridge.Builder.

Update CCMConfig dse option to behave such that if never specified is
not set to false but rather null so CCMBridge.Builder can differentiate
between being expicitly set to false or unset.

Annotate some DSE tests with @DseVersion that didn't previously have it.

Also adds version mappings for DSE 4.7.7-9, 4.8.4-11, and 5.0.1-4.

Result:

withoutDSE option added to indicate CCM configuration should not require
DSE.