Wait for binary port on ccm start
Motivation:

CCM checks for certain messages in the cassandra log, such as "Listening
for thrift clients..." on < 2.2 clusters.  It awaits this message or for
the start process to die before proceeding.  In our setup, we disable
thrift explicitly since we don't use it.

This works fine for Apache Cassandra, but not for DataStax Enterprise,
as the start process remains running indefinitely.

Modifications:

Instead of changing CCM, pass the --no-wait flag to CCM and wait for the
binary protocol port to be listening in CCMBridge as a condition to
exiting start.

Result:

DSE clusters less than 5.0 work with the test harness.