Rename mapper and extras submodules
We don't publish those artifacts currently (it's not clear yet if we'll
need DSE-specific versions).

However, we have to use different names than OSS to avoid a conflict on
`mvn release:prepare`: the OSGI tests depend on the OSS versions, so if
we use the same names the release plugin tries to update those
dependencies, and fails.