JAVA-1358: Add getCenter/getRadius to Distance (#49)
Motivation:

Distance encompasses a center and radius.  Previously only the center
was accessible via its underlying OGC point with getOgcGeometry.

For convenience, accessors for center and radius would be useful for
inspection of this type.

Modifications:

Added getCenter and getRadius methods.

For correctness changed Distance to extend Geometry instead of Point
as extending Point did not provide value.

Override getOgcGeometry to throw UnsupportedOperationException as
previously the Point returned didn't completely represent the Distance
and there is no OGC representation of Distance.

Add javadoc and tests.

Result:

getCenter and getRadius methods added to Distance.