JAVA-1375: Improve error management for continuous queries (#27)
Motivation:

The initial continuous paging implementation is too conservative about
errors: it only releases the stream id when a request has successfully
completed. For errors that originate from the server, we can actually
free the stream id as well, because the server implementation guarantees
that it won't send any more responses.

Modifications:

Track the origin of errors (new `fromServer` parameter in
ContinuousPagingQueue and MultiResponseRequestHandler). On a server
error, release the stream id immediately and don't try to send a cancel
request.
Also, reacquire a connection when we send the cancel request (it uses a
different stream id).

Result:

Server-side errors are returned to the client earlier, and the stream id
is released.
This improved the run time of some server-side tests (see APOLLO-357).

Tests contributed by @tolbertam.