JAVA-1390: Preserve original error when a continuous paging query times out on the client (#31)
Motivation:

If a continuous query fails on the client (ex: timeout waiting for
server pages), we cancel the pending future before trying to complete it
with the error. The client gets a CancellationException instead of the
original error.

Modifications:

Don't cancel the pending future when we already have an error to
complete it with.

Result:

Future is completed with the original error instead of
CancellationException.