Integration tests for DSP-12673 and DSP-12869 (#45)
- Test search predicates against all predicate types

- Add CCMAccess#dsetool and reloadCore and use in tests

Adds dsetool and reloadCore methods.  reloadCore is used for forcing a
solr reindex after inserting data as there is no assurance that solr is
indexed when loading data.  This is used in favor of sleeping for an
arbitrary amount of time.

reloadCore was created as a separate method so it can be used in
ReadOnlyCCMAccess as it is not considered an operation that alters the
state of the cluster.